fix: 修复NumberBox数据绑定类型转换错误
## 修复的问题
- NumberBox.Value属性是double类型，但分数属性是decimal类型
- 导致数据绑定时出现类型转换错误：
  * 'Converter failed to convert value of type '10.0' to type 'Double''
  * 'Cannot save value from target back to source'

## 技术修复
- 将CodeBlank.Score从decimal改为double类型
- 将Question.CSharpDirectScore从decimal改为double类型
- 将Question.TotalScore返回类型从decimal改为double类型
- 将ExamModule.TotalScore返回类型从decimal改为double类型
- 更新StringFormatConverter中的totalScore变量类型
- 在ExamMappingService中添加显式类型转换

## 影响范围
- C#模块的分数输入控件现在可以正常进行双向数据绑定
- 总分计算和显示功能正常工作
- 其他模块的操作点分数仍使用decimal类型（通过显式转换兼容）

## 验证结果
✅ 编译成功，无类型转换错误
✅ NumberBox控件可以正常绑定分数属性
✅ 分数修改时可以正确保存到模型
✅ 总分计算逻辑正常工作

+++ ExamLab/Converters/StringFormatConverter.cs
@@ -154,7 +154,7 @@ public class ModuleTotalScoreConverter : IValueConverter
     {
         if (value is ExamModule module)
         {
-            decimal totalScore = module.Questions.Sum(q => q.TotalScore);
+            double totalScore = module.Questions.Sum(q => q.TotalScore);
             return totalScore;
         }
         return 0.0;
